Fix title merging with next line
In the title for "La causa del accidente que provocó el incendio",
the text is so long that it wraps. This makes the bottom of the "p"
in "provocó" and the dot of the "i" in "incendio" use the same space,
essentially merging both letters.

Just adding a bit of line height fixes it.